+++ bricks/riverpod_simple_architecture/__brick__/lib/bootstrap.dart
@@ -5,11 +5,7 @@ import 'dart:developer';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:hive_flutter/hive_flutter.dart';
 import 'package:platform_info/platform_info.dart';
-import 'package:{{project_name.snakeCase()}}/core/local_storage/app_storage_pod.dart';
-import 'package:{{project_name.snakeCase()}}/init.dart';
-import 'package:{{project_name.snakeCase()}}/shared/riverpod_ext/riverpod_observer.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 // coverage:ignore-file
@@ -38,31 +34,15 @@ final talker = TalkerFlutter.init(
 ///or you can put a provider container in parent
 Future<void> bootstrap(
   FutureOr<Widget> Function() builder, {
-  List<Override> overrides = const [],
-  List<ProviderObserver>? observers,
-  ProviderContainer? parent,
+  required ProviderContainer parent,
 }) async {
   FlutterError.onError = (details) {
     log(details.exceptionAsString(), stackTrace: details.stack);
   };
-  WidgetsFlutterBinding.ensureInitialized();
-  unawaited(init());
-  await Hive.initFlutter();
-  final appBox = await Hive.openBox('appBox');
 
   runApp(
-    ProviderScope(
-      overrides: [
-        appBoxProvider.overrideWithValue(appBox),
-        ...overrides,
-      ],
-      observers: [
-        MyObserverLogger(
-          talker: talker,
-        ),
-        ...?observers,
-      ],
-      parent: parent,
+    UncontrolledProviderScope(
+      container: parent,
       child: await builder(),
     ),
   );