+++ bricks/riverpod_simple_architecture/__brick__/lib/main.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:{{project_name.snakeCase()}}/app/app.dart';
-import 'package:{{project_name.snakeCase()}}/bootstrap.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:{{project_name.snakeCase()}}/splasher.dart';
 
 /// This entry point should be used for production only
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-
-  ///You can override your environment variable in bootstrap method here for providers
-  bootstrap(() => const App());
+  runApp(
+    const ProviderScope(child: Splasher()),
+  );
 }