+++ bricks/riverpod_simple_architecture/__brick__/lib/main_development.dart
@@ -1,8 +1,6 @@
-import 'package:{{project_name.snakeCase()}}/app/app.dart';
-import 'package:{{project_name.snakeCase()}}/bootstrap.dart';
 
 /// This entry point should be used for development only
 void main() {
   ///You can override your environment variable in bootstrap method here for providers
-  bootstrap(() => const App());
+  // bootstrap(() => const App());
 }