+++ bricks/riverpod_simple_architecture/__brick__/lib/splasher.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:{{project_name.snakeCase()}}/app/view/app.dart';
+import 'package:{{project_name.snakeCase()}}/bootstrap.dart';
+import 'package:{{project_name.snakeCase()}}/features/splash/view/splash_view.dart';
+
+class Splasher extends StatelessWidget {
+  const Splasher({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(primaryColor: Colors.blue),
+      home: SplashView(
+        removeSpalshLoader: false,
+        onInitialized: (container) {
+          bootstrap(
+            () => const App(),
+            parent: container,
+          );
+        },
+      ),
+    );
+  }
+}