+++ bricks/riverpod_simple_architecture/__brick__/test/shared/widgets/no_interenet_widget_test.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use, avoid_public_notifier_properties
 
 import 'dart:async';